% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r, R/nog2go.r
\docType{package}
\name{nog2go}
\alias{nog2go}
\alias{nog2go-package}
\title{Function to return the GO category codes for a given NOG term.}
\usage{
nog2go(nog)
}
\arguments{
\item{nog}{A character vector with the NOG terms to look up.}
}
\value{
Either a vector of GO category codes, or a list of vectors if multiple NOG terms were given.
}
\description{
The package \pkg{nog2go} provides a mapping from NOG terms to GO category codes.
}
\details{
If there are no codes for a given term, the function will return character(0).
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>

Chris Field <fieldc@ethz.ch">
}
\keyword{None}
